pyproject.toml
[tool.poetry]
name = "hw-11-web"
version = "0.1.0"
description = ""
authors = ["SergRusetskiy1974 <113336060+rusetskiy1974@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.13.1"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = {extras = ["email"], version = "^2.6.4"}
libgravatar = "^1.0.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
fastapi-mail = "^1.4.1"
pydantic-settings = "^2.2.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.39.1"
bcrypt = "4.0.1"
redis-lru = "^0.1.2"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
pytest = "^8.1.1"
pytest-mock = "^3.14.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"


[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "tests",
]
pythonpath = "."
filterwarnings = [
    "ignore::DeprecationWarning",
]
